rename vc_tag and vc_changes
