+++ services/web/server/src/simcore_service_webserver/meta_modeling_version_control.py
@@ -11,11 +11,11 @@ from models_library.utils.fastapi_encoders import jsonable_encoder
 
 from .projects.project_models import ProjectDict
 from .version_control.errors import UserUndefined
-from .version_control.vctags import compose_workcopy_project_tag_name
-from .version_control.version_control_changes import (
+from .version_control.vc_changes import (
     compute_workbench_checksum,
     eval_workcopy_project_id,
 )
+from .version_control.vc_tags import compose_workcopy_project_tag_name
 from .version_control.version_control_db import VersionControlRepository
 from .version_control.version_control_models import CommitID, TagProxy
