+++ services/web/server/src/simcore_service_webserver/version_control/version_control_db.py
@@ -31,8 +31,8 @@ from .errors import (
     NoCommitError,
     NotFoundError,
 )
-from .vctags import parse_workcopy_project_tag_name
-from .version_control_changes import compute_workbench_checksum
+from .vc_changes import compute_workbench_checksum
+from .vc_tags import parse_workcopy_project_tag_name
 from .version_control_models import (
     HEAD,
     CommitID,
